Eject from shared workflows (#2)
The template repository from which this repository was generated uses
shared workflows[^1] on GitHub. This approach had the advantage of
quickly bootstrapping new repositories, but introduced challenges later
on when workflows needed to be customized. FlowCrafter, i.e. this tool,
is the solution to this problem. The repository has been "ejected" from
the shared workflows with the goal to eventually use FlowCrafter to
manage its workflows.

[^1]: https://docs.github.com/en/actions/using-workflows/reusing-workflows