Create model for jobs
A simple struct has been created that represents a job in a workflow
file. Each job has a unique name and a template that contains the job
definition.

Jobs can be constructed using a builder pattern. The builder returns a
`Result`, and errors out if either `name` or `template` are missing.