Create model for workflows
A simple struct has been created that represents a workflow in a
workflow file. Each workflow has a unique name, a template that contains
the workflow definition, and a list of jobs that belong to this
workflow.

Workflows can be constructed using a builder pattern. The builder
returns a `Result`, and errors out if either `name` or `template` are
missing. A workflow without a job is theoretically valid and will not
result in an error.