Simplify error type (#6)
It is too early in the development cycle of the crate to design a robust
error type. So we will use a single variant and anyhow for now to
generate meaningful error messages. As we start integrating the crate
into other applications, we will learn what the needs of consumers are
and can refactor the error type.