Create command to initialize repository (#10)
The command to initialize a new repository has been finalized. The
command takes the name of a GitHub repository as an argument and writes
it to a configuration file inside the repositories `.github` directory.
Subsequent executions of the flowcrafter CLI can then read this
configuration and save the user from providing the settings on each
invocation of the command-line tool.