Replace builders with derive macro
The builder pattern[^1] is very common in Rust and has been used in
this crate as well to initialize types. The implementations of the
different builders are all very similar and can be replaced with a
derive macro.

We are using the typed-builder[^2], which offers compile-time
verification that all fields are set exactly once.

[^1]: https://rust-unofficial.github.io/patterns/patterns/creational/builder.html
[^2]: https://crates.io/crates/typed-builder