Use anyhow error in command-line application (#20)
The custom error type that was created for the command-line interface of
FlowCrafter has been removed in favor of the generic error provided by
the anyhow[^1] crate. Custom errors are great for libraries, where
consumers might want to handle errors in different ways. For binaries,
anyhow is a better fit, since it provides more readable error messages
with a lot more context to the user.

[^1]: https://crates.io/crates/anyhow