Ignore newlines when decoding base64 content
The base64 crate for Rust (correctly) treats the newline character `\n`
as an invalid byte in the base64 alphabet[^1]. This causes an issue when
decoding a file's content from GitHub, which breaks the base64 string
into lines of 60 characters.

When decoding a file from GitHub, newlines are thus removed before
trying to decode the base64 string.

[^1]: https://github.com/marshallpierce/rust-base64/issues/172