Create custom renderer (#24)
The previous attempt to render the templates with the Liquid templating
engine has failed. For one, GitHub uses as delimiters for `${{` and `}}`
to wrap variables, which collides with Liquid's syntax. For another,
Liquid does not support indenting multiline strings. Both problems exist
with other popular templating engines as well, which is why we have
decided to implement our own solution.

The initial implementation of FlowCrafter's render is very naive and
simply appends the jobs to the end of the workflow.