Create temporary project for tests (#32)
A new struct has been created to make it easier to set up a project for
unit tests. The new struct creates a temporary directory and sets up the
necessary folder structure to initialize a project in it. Existing tests
have been rewritten to use this struct, which removes double test setup
code.