Build release artifacts with cargo-dist
cargo-dist[^1] is a new tool that makes it easy to build and package
release artifacts for Rust applications. It supports integration with
GitHub, where it builds artifacts on GitHub Actions and publishes them
to a matching GitHub release.

We are somewhat opinionated about our release process, and have tweaked
the default configuration for cargo-dist to match our existing process.
Releases are still prepared and published using the GitHub UI, and
cargo-dist will only handle the building of release artifacts.

[^1]: https://github.com/axodotdev/cargo-dist