Remove GitHub Action for FlowCrafter (#118)
As it turns out, GitHub Actions cannot update workflow files. The
necessary permission can only be granted to GitHub Apps, and the only
other solution is to pass a Personal Access Token (PAT) to the action.
The latter is difficult to manage at scale, especially with the new
granular tokens that must expire within a year. Until a better solution
can be found, we are removing the GitHub Action again.

This reverts commit 6d68f5e84974194896fcba1c1645c0b2bb6a7f9d.